requirements.txt
ollama
streamlit
yfinance
pandas
mysqlclient
python-dotenv